+++ src/lib/envs/client.ts
@@ -1,6 +1,6 @@
 import {
   PUBLIC_ALLOWED_REQUEST_METHODS,
-  PUBLIC_CSRF_IGNORE_PATHS,
+  PUBLIC_CSRF_SKIP_ROUTES,
   PUBLIC_SESSION_KEY
 } from '$env/static/public'
 
@@ -8,8 +8,8 @@ export const allowedRequestMethods = PUBLIC_ALLOWED_REQUEST_METHODS.split(',')
   .map(method => method.trim())
   .filter(method => method)
 
-export const csrfIgnorePaths = PUBLIC_CSRF_IGNORE_PATHS.split(',')
-  .map(path => path.trim())
-  .filter(path => path)
+export const csrfSkipRoutes = PUBLIC_CSRF_SKIP_ROUTES.split(',')
+  .map(route => route.trim())
+  .filter(route => route)
 
 export const sessionKey = PUBLIC_SESSION_KEY