+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+
+### Fixed
+- Fix `allowedRequestMethods` returning array with single empty string when env var not set
+
 ## [0.2.3] - 2022-09-17
 
 ### Fixed