+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased] - 
 
 ### Changed
-- Upgrade SvelteKit to v1.0.0-next.589
+- Upgrade SvelteKit to v1.0
 
 ## [0.5.0] - 2022-11-07